requirements.txt
aiohttp>=3.8.0
pyyaml>=6.0
lxml>=4.9.0
